Refactor uptime-kuma to match AGENTS.md structure
Split the two largest uptime-kuma source files into one-concept-per-file
modules, following the ledger/ and db/attendees/ reference patterns:

client.ts (398 lines) becomes:
- authorization.ts (56) — reading the effective Authorization header
- version.ts (98) — capturing and checking Kuma's post-login version
- schemas.ts (116) — valibot schemas for monitor-list and responses
- event-capture.ts (94) — timing out socket events and acks
- client.ts (140) — thin shell: client factory, connect, socket URL

monitors.ts (342 lines) becomes:
- errors.ts (34) — mapping typed Kuma failures to catalog copy
- matching.ts (210) — pure predicates for finding a site's monitor
- monitors.ts (230) — reconciliation orchestration and service entry

Schema-tize monitor matching: the nine checks that decide whether a Kuma
monitor can drive a site's scheduled request are now a declarative
SCHEDULED_MONITOR_RULES table folded over with every(), so adding a new
rule is one line in the table, not a new arm on an if chain.

Deduplicate the version timeout: version.ts no longer has its own
withVersionTimeout copy of the Promise.race+setTimeout pattern; it
calls withEventTimeout from event-capture.ts.

Adopt #fp pipe+filter for the shared-groups predicate.

All uptime-kuma files are now under 400 lines. Full precommit passes.
