+++ src/shared/uptime-kuma/client.ts
@@ -1,38 +1,24 @@
-import * as v from "valibot";
-import { bearerAuthorization, bearerTokenOrNull } from "#shared/bearer.ts";
 import { countExternalSubrequest } from "#shared/subrequest-budget.ts";
-import { integerAtLeast } from "#shared/validation/number.ts";
-import { parseOrThrow } from "#shared/validation/parse.ts";
 import type { UptimeKumaConfig } from "./config.ts";
-import { UptimeKumaError, type UptimeKumaErrorKind } from "./error.ts";
+import { UptimeKumaError } from "./error.ts";
+import { call, callMonitorList, monitorListCapture } from "./event-capture.ts";
+import {
+  AddResponseSchema,
+  LoginResponseSchema,
+  MonitorListSchema,
+  parseKumaResponse,
+  requireOk,
+  type UptimeKumaMonitor,
+  type UptimeKumaMonitorInput,
+} from "./schemas.ts";
 import {
   createUptimeKumaSocket,
   type UptimeKumaSocket,
   uptimeKumaConnectionError,
 } from "./socket.ts";
+import { captureVersion } from "./version.ts";
 
-const SOCKET_TIMEOUT_MS = 10_000;
-const MONITOR_LIST_TIMEOUT_MS = 60_000;
-
-type SocketListener = (...args: unknown[]) => void;
-
-export type UptimeKumaMonitorInput = Record<string, unknown>;
-
-export interface UptimeKumaMonitor {
-  acceptedStatusCodes: string[];
-  active: boolean;
-  authorization: string | null;
-  conditions: unknown[];
-  id: number;
-  interval: number;
-  method: string;
-  name: string;
-  parent: number | null;
-  timeout: number;
-  type: string;
-  upsideDown: boolean;
-  url: string | null;
-}
+export type { UptimeKumaMonitor, UptimeKumaMonitorInput };
 
 export interface UptimeKumaClient {
   addMonitor(monitor: UptimeKumaMonitorInput): Promise<number>;
@@ -42,249 +28,9 @@ export interface UptimeKumaClient {
   login(username: string, password: string): Promise<void>;
 }
 
-const ActiveSchema = v.pipe(
-  v.union([v.boolean(), v.literal(0), v.literal(1)]),
-  v.transform((value) => value === true || value === 1),
-);
+type SocketListener = (...args: unknown[]) => void;
 
-const CustomHeadersSchema = v.record(v.string(), v.string());
-
-type CustomAuthorization = {
-  authorization: string | null;
-  valid: boolean;
-};
-
-const readCustomAuthorization = (
-  headers: string | null,
-): CustomAuthorization => {
-  if (headers === null || headers === "")
-    return { authorization: null, valid: true };
-  try {
-    const values = v.parse(CustomHeadersSchema, JSON.parse(headers));
-    const entry = Object.entries(values).find(
-      ([name]) => name.toLowerCase() === "authorization",
-    );
-    return {
-      authorization: entry === undefined ? null : entry[1],
-      valid: true,
-    };
-  } catch {
-    // A malformed custom header belongs to a different, broken monitor.
-    return { authorization: null, valid: false };
-  }
-};
-
-const authorizationFor = (
-  headers: string | null,
-  authMethod: string | null,
-  bearerToken: string | null,
-): string | null => {
-  const custom = readCustomAuthorization(headers);
-  if (!custom.valid) return null;
-  if (custom.authorization !== null) {
-    const token = bearerTokenOrNull(custom.authorization);
-    return token === null ? custom.authorization : bearerAuthorization(token);
-  }
-  return authMethod === "bearer" && bearerToken !== null
-    ? bearerAuthorization(bearerToken)
-    : null;
-};
-
-const RawMonitorSchema = v.object({
-  accepted_statuscodes: v.array(v.string()),
-  active: ActiveSchema,
-  authMethod: v.nullable(v.string()),
-  bearer_token: v.nullable(v.string()),
-  conditions: v.array(v.unknown()),
-  headers: v.nullable(v.string()),
-  id: integerAtLeast(1),
-  interval: integerAtLeast(1),
-  method: v.string(),
-  name: v.string(),
-  parent: v.nullable(integerAtLeast(1)),
-  timeout: integerAtLeast(1),
-  type: v.string(),
-  upsideDown: ActiveSchema,
-  url: v.nullable(v.string()),
-});
-const MonitorSchema = v.pipe(
-  RawMonitorSchema,
-  v.transform(
-    ({
-      accepted_statuscodes,
-      authMethod,
-      bearer_token,
-      headers,
-      ...monitor
-    }) => ({
-      ...monitor,
-      acceptedStatusCodes: accepted_statuscodes,
-      authorization: authorizationFor(headers, authMethod, bearer_token),
-    }),
-  ),
-);
-
-const MonitorListSchema = v.record(v.string(), MonitorSchema);
-const VersionNumberSchema = v.pipe(
-  v.string(),
-  v.regex(/^\d+$/),
-  v.transform(Number),
-  v.safeInteger(),
-);
-const VersionSchema = v.pipe(
-  v.string(),
-  v.transform((version) => version.split(".")),
-  v.tuple([VersionNumberSchema, VersionNumberSchema]),
-);
-const InfoSchema = v.object({ version: v.optional(VersionSchema) });
-const FailedResponseSchema = v.object({
-  msg: v.string(),
-  ok: v.literal(false),
-});
-const PlainLoginFailureSchema = v.object({
-  msg: v.string(),
-  msgi18n: v.optional(v.literal(false)),
-  ok: v.literal(false),
-});
-const IncorrectCredentialsSchema = v.object({
-  msg: v.literal("authIncorrectCreds"),
-  msgi18n: v.literal(true),
-  ok: v.literal(false),
-});
-const OkResponseSchema = v.object({ ok: v.literal(true) });
-const BasicResponseSchema = v.union([OkResponseSchema, FailedResponseSchema]);
-const LoginResponseSchema = v.union([
-  OkResponseSchema,
-  IncorrectCredentialsSchema,
-  PlainLoginFailureSchema,
-  v.object({ tokenRequired: v.literal(true) }),
-]);
-const AddResponseSchema = v.union([
-  v.object({ monitorID: integerAtLeast(1), ok: v.literal(true) }),
-  FailedResponseSchema,
-]);
-
-const parseKumaResponse = <TSchema extends v.GenericSchema>(
-  schema: TSchema,
-  value: unknown,
-): v.InferOutput<TSchema> =>
-  parseOrThrow(schema, value, () => new UptimeKumaError("invalid_response"));
-
-const requireOk = (value: unknown): void => {
-  const response = parseKumaResponse(BasicResponseSchema, value);
-  if (!response.ok) throw new Error(response.msg);
-};
-
-type EventCapture = {
-  cancel: () => void;
-  latest: () => unknown;
-  received: Promise<void>;
-};
-
-type VersionOutcome = { error: unknown; ok: false } | { ok: true };
-
-type VersionCapture = {
-  cancel: () => void;
-  read: () => Promise<void>;
-};
-
-const requireSupportedVersion = ([major, minor]: [number, number]): void => {
-  if (major < 2 || (major === 2 && minor < 4)) {
-    throw new UptimeKumaError("unsupported_version");
-  }
-};
-
-type TimedEvent = "info" | "monitorList";
-
-const TIMEOUT_ERROR_KINDS: Record<TimedEvent, UptimeKumaErrorKind> = {
-  info: "version_timeout",
-  monitorList: "monitor_list_timeout",
-};
-
-const withEventTimeout = async <Value>(
-  promise: Promise<Value>,
-  event: TimedEvent,
-  timeoutMs = SOCKET_TIMEOUT_MS,
-): Promise<Value> => {
-  const timeout = Promise.withResolvers<never>();
-  const timer = setTimeout(() => {
-    timeout.reject(new UptimeKumaError(TIMEOUT_ERROR_KINDS[event]));
-  }, timeoutMs);
-  try {
-    return await Promise.race([promise, timeout.promise]);
-  } finally {
-    clearTimeout(timer);
-  }
-};
-
-const captureVersion = (socket: UptimeKumaSocket): VersionCapture => {
-  let listener: SocketListener;
-  const promise = new Promise<VersionOutcome>((resolve) => {
-    listener = (value) => {
-      try {
-        const parsed = v.safeParse(InfoSchema, value);
-        if (!parsed.success) {
-          throw new UptimeKumaError("unsupported_version");
-        }
-        const info = parsed.output;
-        if (info.version === undefined) {
-          socket.once("info", listener);
-          return;
-        }
-        requireSupportedVersion(info.version);
-        resolve({ ok: true });
-      } catch (error) {
-        resolve({ error, ok: false });
-      }
-    };
-    socket.once("info", listener);
-  });
-  return {
-    cancel: () => socket.off("info", listener),
-    read: async (): Promise<void> => {
-      const outcome = await withEventTimeout(promise, "info");
-      if (!outcome.ok) throw outcome.error;
-    },
-  };
-};
-
-const captureEvents = (
-  socket: UptimeKumaSocket,
-  event: TimedEvent,
-  timeoutMs: number,
-): EventCapture => {
-  let latest: unknown;
-  let listener: SocketListener;
-  const received = Promise.withResolvers<void>();
-  listener = (value) => {
-    latest = value;
-    received.resolve();
-    socket.once(event, listener);
-  };
-  socket.once(event, listener);
-  return {
-    cancel: () => {
-      socket.off(event, listener);
-      received.resolve();
-    },
-    latest: () => latest,
-    received: withEventTimeout(received.promise, event, timeoutMs),
-  };
-};
-
-const callWithTimeout = (
-  socket: UptimeKumaSocket,
-  event: string,
-  timeoutMs: number,
-  ...args: unknown[]
-): Promise<unknown> => socket.timeout(timeoutMs).emitWithAck(event, ...args);
-
-const call = (
-  socket: UptimeKumaSocket,
-  event: string,
-  ...args: unknown[]
-): Promise<unknown> =>
-  callWithTimeout(socket, event, SOCKET_TIMEOUT_MS, ...args);
+const SOCKET_TIMEOUT_MS = 10_000;
 
 export const createUptimeKumaClient = (
   socket: UptimeKumaSocket,
@@ -307,14 +53,10 @@ export const createUptimeKumaClient = (
       socket.disconnect();
     },
     getMonitors: async (): Promise<UptimeKumaMonitor[]> => {
-      const list = captureEvents(
-        socket,
-        "monitorList",
-        MONITOR_LIST_TIMEOUT_MS,
-      );
+      const list = monitorListCapture(socket);
       try {
         const [response] = await Promise.all([
-          callWithTimeout(socket, "getMonitorList", MONITOR_LIST_TIMEOUT_MS),
+          callMonitorList(socket),
           list.received,
         ]);
         requireOk(response);
